requirements.txt
langchain
openai
tiktoken
faiss_cpu
sentence-transformers
python-dotenv
